Ændring til reader.

+++ REST Service/DBUtil/FaerdigVareManager.cs
@@ -59,9 +59,9 @@ namespace REST_Service.DBUtil
 
             tempFaerdigVare.FaerdigVare_Nr = reader.GetInt32(0);
             tempFaerdigVare.FaerdigVareNavn = reader.GetString(1);
-            tempFaerdigVare.Min = reader.GetInt32(2);
-            tempFaerdigVare.Max = reader.GetInt32(3);
-            tempFaerdigVare.Snit = reader.GetInt32(4);
+            tempFaerdigVare.Min = reader.GetDouble(2);
+            tempFaerdigVare.Max = reader.GetDouble(3);
+            tempFaerdigVare.Snit = reader.GetDouble(4);
 
             return tempFaerdigVare;
         }
